Move labels parser to modules, remove unused prop from selection dialog
